Model/requirements.txt
fastapi
uvicorn
torch
safetensors
tokenizers
